+++ packages/shiki/src/__tests__/custom-language/rockstar.tmLanguage.json
